+++ frontend/Active-learning-project/src/app/app.module.ts
@@ -8,9 +8,8 @@ import { MasterLayoutModule } from './master-layout/master-layout.module';
 import { SocialLoginModule, GoogleLoginProvider } from '@abacritt/angularx-social-login';
 import { environment } from 'src/environments/environment';
 import { SigninAuthService } from './shared/helpers/signin-auth.service';
-import { AuthenticateService } from './user-authentication/services/authenticate.service';
 import { ErrorsInterceptor } from './shared/helpers/errors.interceptor';
-import { APP_CONFIG, APP_SERVICE_CONFIG } from './app-config/app-config.service';
+import { APP_CONFIG, APP_SERVICE_CONFIG } from './services/app-config/app-config.service';
 import { FullNamePipe } from './shared/pipes/full-name.pipe';
 import { AccountComponent } from './user-authentication/account/account.component';
 import { GithubButtonComponent } from './user-authentication/github-button/github-button.component';
@@ -19,65 +18,65 @@ import { SignInComponent } from './user-authentication/signin/signin.component';
 import { SignupComponent } from './user-authentication/signup/signup.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { AppRoutingModule } from './app.module.routing';
+import { AuthenticateService } from './services/user-authentication/authenticate.service';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    FullNamePipe,
-    AccountComponent,
-    SignInComponent,
-    SignupComponent,
-    GoogleButtonComponent,
-    GithubButtonComponent
-  ],
-  imports: [
-    BrowserModule,
-    BrowserAnimationsModule,
-    AppRoutingModule,
-    MasterLayoutModule,
-    SocialLoginModule,
-    HttpClientModule,
-    SocialLoginModule,
-    FormsModule,
-    ReactiveFormsModule,
-    ToastrModule.forRoot({
-      positionClass: 'toast-bottom-right',
-      preventDuplicates: true,
-      progressBar: true,
-    })
-  ],
-  providers: [
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: SigninAuthService,
-      multi: true,
-    },
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: ErrorsInterceptor,
-      multi: true,
-    },
-    {
-      provide: APP_SERVICE_CONFIG,
-      useValue: APP_CONFIG,
-    },
-    {
-      provide: 'SocialAuthServiceConfig',
-      useValue: {
-        autoLogin: false,
-        providers: [
-          {
-            id: GoogleLoginProvider.PROVIDER_ID,
-            provider: new GoogleLoginProvider(environment.googleAuthClientID),
-          },
-        ],
-        onError: (err: any) => {
-          console.error(err);
-        },
-      },
-    },
-    AuthenticateService,
-  ],
-  bootstrap: [AppComponent],
+	declarations: [
+		AppComponent,
+		FullNamePipe,
+		AccountComponent,
+		SignInComponent,
+		SignupComponent,
+		GoogleButtonComponent,
+		GithubButtonComponent
+	],
+	imports: [
+		BrowserModule,
+		BrowserAnimationsModule,
+		AppRoutingModule,
+		MasterLayoutModule,
+		SocialLoginModule,
+		HttpClientModule,
+		FormsModule,
+		ReactiveFormsModule,
+		ToastrModule.forRoot({
+			positionClass: 'toast-bottom-right',
+			preventDuplicates: true,
+			progressBar: true,
+		})
+	],
+	providers: [
+		AuthenticateService,
+		{
+			provide: HTTP_INTERCEPTORS,
+			useClass: SigninAuthService,
+			multi: true,
+		},
+		{
+			provide: HTTP_INTERCEPTORS,
+			useClass: ErrorsInterceptor,
+			multi: true,
+		},
+		{
+			provide: APP_SERVICE_CONFIG,
+			useValue: APP_CONFIG,
+		},
+		{
+			provide: 'SocialAuthServiceConfig',
+			useValue: {
+				autoLogin: false,
+				providers: [
+					{
+						id: GoogleLoginProvider.PROVIDER_ID,
+						provider: new GoogleLoginProvider(environment.googleAuthClientID),
+					},
+				],
+				onError: (err: any) => {
+					console.error(err);
+				},
+			},
+		}
+	],
+	bootstrap: [AppComponent],
 })
 export class AppModule { }
