+++ frontend/Active-learning-project/src/app/services/app-config/app-config-bundle.service.ts
@@ -1,10 +1,9 @@
 import { Inject, Injectable } from '@angular/core';
 import { AppConfig } from './app-config.interface';
 import { APP_CONFIG } from './app-config.service';
-import { AccountModule } from '../user-authentication/account.module';
 
 @Injectable({
-  providedIn: AccountModule
+  providedIn: "root"
 })
 export class AppConfigBundleService {
   constructor(@Inject(APP_CONFIG) private config: AppConfig) { }
