+++ frontend/Active-learning-project/src/app/services/app-config/app-config.interface.ts
