+++ frontend/Active-learning-project/src/app/services/user-authentication/authenticate.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { AuthenticateService } from './authenticate.service';
+
+describe('AuthenticateService', () => {
+  let service: AuthenticateService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(AuthenticateService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
