+++ frontend/Active-learning-project/src/app/services/user-management/user-management.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { UserManagementService } from './user-management.service';
+
+describe('UserManagementService', () => {
+  let service: UserManagementService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(UserManagementService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
