+++ frontend/Active-learning-project/src/app/user-authentication/github-button/github-button.component.ts
@@ -4,9 +4,9 @@ import { ToastrService } from 'ngx-toastr';
 import queryString from 'query-string';
 import { Observable, Observer } from 'rxjs';
 import { UserRequest } from 'src/app/models/payloads/requests/user.auth.request.model';
+import { AuthenticateService } from 'src/app/services/user-authentication/authenticate.service';
+import { GithubRequestService } from 'src/app/services/user-authentication/github-request.service';
 import { auth } from 'src/app/shared/routers/Routers';
-import { AuthenticateService } from '../services/authenticate.service';
-import { GithubRequestService } from '../services/github-request.service';
 
 @Component({
   selector: 'alp-github-button',
