+++ frontend/Active-learning-project/src/app/user-authentication/signin/signin.component.html
@@ -81,17 +81,13 @@
             Invalid Login Credentials
           </span>
         </button>
-        <a
-          mat-button
-          class="menu-button"
-          class="btn btn-forgot text-dark"
-        >
+        <a mat-button class="menu-button" class="btn btn-forgot text-dark">
           Forgot Password?
         </a>
       </div>
     </form>
     <div class="auth-signup-link text-center">
-      <i>Don't have an account? <a [routerLink]="['/signup']">Sign Up</a></i>
+      <i>Don't have an account? <a href="/signup">Sign Up</a></i>
     </div>
   </div>
 </div>
